this can launch processccd inside the python process using the commandline runner with args explicitly specified rather than coming from sys.
it will, however, definitely fail at the end because that command
line tool does a system exit (which is an exception)

I'm not handling that because I want to move on from this particular
invocation form anyway.

This also disables the big download/prep stage because I want to
test processCcd invocations without taking 20 minutes each time
for that.